Separate routes and controllers, Modal size
