+++ views/edit.ejs
@@ -1,6 +1,6 @@
    <div class="modal fade" id="example-<%=contact.id%>" tabindex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true">
         
-        <div class="modal-dialog modal-dialog-centered" role="document">
+        <div class="modal-dialog modal-dialog-scrollable modal-lg" role="document">
           <div class="modal-content">
             <div class="modal-header">
               <h5 class="modal-title" id="exampleModalLongTitle">Edit contact</h5>
@@ -10,20 +10,30 @@
             </div>
             <div class="modal-body">
               <form id="edit-contact-<%=contact.id%>" action="/func/edit" method="Post" class="row g-3 needs-validation" novalidate>
-                <div class="col-md-6">
+                <div class="row">
+                <div class="col-md-4">
                     <label for="validationCustom01" class="form-label" >First name*</label>
                     <input type="text" class="form-control" name="first_name" id="validationCustom01" value="<%=contact.firstname%>"  required>
                     <div class="valid-feedback">
                         Looks good!
                     </div>
                 </div>
-                <div class="col-md-6">
+                <div class="col-md-4">
                     <label for="validationCustom02" class="form-label">Last name*</label>
                     <input type="text" class="form-control" name="last_name" id="validationCustom02" value="<%=contact.lastname%>" required>
                     <div class="valid-feedback">
                         Looks good!
                     </div>
                 </div>
+                <div class="col-md-4">
+                    <label class="form-label">Phone Number*</label>
+                    <input type="number" value="<%=contact.phoneNumber%>" name="phone_number" class="form-control" placeholder="1234567898" required>
+                    <div class="valid-feedback">
+                        Looks good!
+                    </div>
+                </div>
+                </div>
+                <div class="row">
                 <div class="col-md-6">
                     <label for="validationCustomUsername" class="form-label">Email*</label>
                     <div class="input-group has-validation">
@@ -35,18 +45,12 @@
                         </div>
                     </div>
                 </div>
+                
                 <div class="col-md-6">
-                    <label class="form-label">Phone Number*</label>
-                    <input type="number" value="<%=contact.phoneNumber%>" name="phone_number" class="form-control" placeholder="1234567898" required>
-                    <div class="valid-feedback">
-                        Looks good!
-                    </div>
-                </div>
-                <div class="col-md-12">
                     <label class="form-label" for="start">Birth Date</label>
                     <input type="date" value="<%=contact.birthdate%>" class="form-control" id="bday" name="bday">
                 </div>
-        
+            </div>
                 <div class="col-12">
                     <label for="inputAddress" class="form-label">Address</label>
                     <input type="text" value="<%=contact.address1%>" class="form-control" name="address_1" id="inputAddress" placeholder="1234 Main St">
@@ -55,14 +59,15 @@
                     <label for="inputAddress2" class="form-label">Address 2</label>
                     <input type="text" value="<%=contact.address2%>"class="form-control" name="address_2" id="inputAddress2" placeholder="Apartment, studio, or floor">
                 </div>
-                <div class="col-md-6">
+                <div class="row">
+                <div class="col-md-4">
                     <label for="validationCustom03" class="form-label">City</label>
                     <input type="text" value="<%=contact.city%>" class="form-control" name="city" id="validationCustom03" required>
                     <div class="invalid-feedback">
                         Please provide a valid city.
                     </div>
                 </div>
-                <div class="col-md-3">
+                <div class="col-md-4">
                     <label for="validationCustom04" class="form-label">State</label>
                     <select class="form-select" name="state" id="validationCustom04" required>
                         <option>Delhi</option>
@@ -75,13 +80,15 @@
                         Please select a valid state.
                     </div>
                 </div>
-                <div class="col-md-3">
+                <div class="col-md-4">
                     <label for="validationCustom05" class="form-label">Zip</label>
                     <input value="<%=contact.zip%>" type="text" class="form-control" name="zip" id="validationCustom05" required>
                     <div class="invalid-feedback">
                         Please provide a valid zip.
                     </div>
                 </div>
+            </div>
+            <br>
                 <div class="col-12">
                     <div class="form-check">
                         <input class="form-check-input" type="checkbox" value="" id="invalidCheck" required>
@@ -94,12 +101,11 @@
                     </div>
                 </div>
                 <input type="hidden" name="contact_id" value="<%=contact._id%>">
-              <button type="submit" id="contact-<%=contact._id%>" onclick="ajaxCode(this)" class="btn btn-primary">Save changes</button>
-
+              
             </form>
             </div>
             <div class="modal-footer">
-              <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                <button type="submit" id="contact-<%=contact._id%>" onclick="ajaxCode(this)" class="btn btn-primary">Save changes</button>
             </div>
           </div>
         </div>
